Switch to posted MMIO by default
Now there is a way to signal to the kernel to map memory as non-posted
we can switch the default device memory to a posted type.

This should help with device driver performance as they don't need to
wait for memory writes to be acknowledged. It will also align us with
32-bit arm that already uses the equivalent memory type.