dev/fdt: Add fdt_foreach_mem_region
This is used early in the boot to find physical memory. Previously it
needed an array of memory regions to be passed in. We keep finding
hardware where this array is too small causing the kernel to smash the
stack.

Replace with a function that takes a callback that can insert the
memory into the physical map directly.

Reviewed by:	imp
Sponsored by:	Arm Ltd
Differential Revision:	https://reviews.freebsd.org/D49697